refactor(account): implement AccountType enum and BigDecimal +

+++ src/main/java/com/jindero/banking/features/account/Chargeable.java
@@ -1,7 +1,9 @@
 package com.jindero.banking.features.account;
 
+import java.math.BigDecimal;
+
 public interface Chargeable {
 
-  double calculateFees();
+  BigDecimal calculateFees();
   void applyMonthlyFee();
 }
